+++ virtualizarr/tests/test_manifests/test_manifest.py
@@ -1,7 +1,6 @@
 import pytest
-from pydantic import ValidationError
 
-from virtualizarr.manifests import ChunkManifest, concat_manifests, stack_manifests
+from virtualizarr.manifests import ChunkManifest
 
 
 class TestCreateManifest:
@@ -25,7 +24,7 @@ class TestCreateManifest:
         chunks = {
             "0.0.0": {"path": "s3://bucket/foo.nc"},
         }
-        with pytest.raises(ValidationError, match="missing"):
+        with pytest.raises(ValueError, match="must be of the form"):
             ChunkManifest(entries=chunks)
 
         chunks = {
@@ -35,7 +34,7 @@ class TestCreateManifest:
                 "length": 100,
             },
         }
-        with pytest.raises(ValidationError, match="should be a valid integer"):
+        with pytest.raises(ValueError, match="must be of the form"):
             ChunkManifest(entries=chunks)
 
     def test_invalid_chunk_keys(self):
@@ -50,20 +49,6 @@ class TestCreateManifest:
             "0": {"path": "s3://bucket/foo.nc", "offset": 200, "length": 100},
         }
         with pytest.raises(ValueError, match="Inconsistent number of dimensions"):
-            ChunkManifest(entries=chunks)
-
-        chunks = {
-            "0.0.0": {"path": "s3://bucket/foo.nc", "offset": 100, "length": 100},
-            "0.0.1": {"path": "s3://bucket/foo.nc", "offset": 200, "length": 100},
-            "0.1.0": {"path": "s3://bucket/foo.nc", "offset": 300, "length": 100},
-        }
-        with pytest.raises(ValueError, match="do not form a complete grid"):
-            ChunkManifest(entries=chunks)
-
-        chunks = {
-            "1": {"path": "s3://bucket/foo.nc", "offset": 100, "length": 100},
-        }
-        with pytest.raises(ValueError, match="do not form a complete grid"):
             ChunkManifest(entries=chunks)
 
 
@@ -94,64 +79,7 @@ class TestEquals:
                 "0.0.1": {"path": "foo.nc", "offset": 400, "length": 100},
             }
         )
-        assert not manifest1 == manifest2
         assert manifest1 != manifest2
-
-
-# TODO could we use hypothesis to test this?
-# Perhaps by testing the property that splitting along a dimension then concatenating the pieces along that dimension should recreate the original manifest?
-class TestCombineManifests:
-    def test_concat(self):
-        manifest1 = ChunkManifest(
-            entries={
-                "0.0.0": {"path": "foo.nc", "offset": 100, "length": 100},
-                "0.0.1": {"path": "foo.nc", "offset": 200, "length": 100},
-            }
-        )
-        manifest2 = ChunkManifest(
-            entries={
-                "0.0.0": {"path": "foo.nc", "offset": 300, "length": 100},
-                "0.0.1": {"path": "foo.nc", "offset": 400, "length": 100},
-            }
-        )
-        axis = 1
-        expected = ChunkManifest(
-            entries={
-                "0.0.0": {"path": "foo.nc", "offset": 100, "length": 100},
-                "0.0.1": {"path": "foo.nc", "offset": 200, "length": 100},
-                "0.1.0": {"path": "foo.nc", "offset": 300, "length": 100},
-                "0.1.1": {"path": "foo.nc", "offset": 400, "length": 100},
-            }
-        )
-
-        result = concat_manifests([manifest1, manifest2], axis=axis)
-        assert result.dict() == expected.dict()
-
-    def test_stack(self):
-        manifest1 = ChunkManifest(
-            entries={
-                "0.0": {"path": "foo.nc", "offset": 100, "length": 100},
-                "0.1": {"path": "foo.nc", "offset": 200, "length": 100},
-            }
-        )
-        manifest2 = ChunkManifest(
-            entries={
-                "0.0": {"path": "foo.nc", "offset": 300, "length": 100},
-                "0.1": {"path": "foo.nc", "offset": 400, "length": 100},
-            }
-        )
-        axis = 1
-        expected = ChunkManifest(
-            entries={
-                "0.0.0": {"path": "foo.nc", "offset": 100, "length": 100},
-                "0.0.1": {"path": "foo.nc", "offset": 200, "length": 100},
-                "0.1.0": {"path": "foo.nc", "offset": 300, "length": 100},
-                "0.1.1": {"path": "foo.nc", "offset": 400, "length": 100},
-            }
-        )
-
-        result = stack_manifests([manifest1, manifest2], axis=axis)
-        assert result.dict() == expected.dict()
 
 
 @pytest.mark.skip(reason="Not implemented")